server/services/auth_service/requirements.txt
fastapi
uvicorn
pika
sqlalchemy
pydantic
asyncpg
passlib[bcrypt,argon2]
email-validator
pyjwt
python-multipart
pydantic_settings